refactored middleware
